DeleteBlog Functionality added. Working on EditBlog

+++ Web/MyHotelWebsite.Web/Areas/Administration/Controllers/BlogsController.cs
@@ -47,7 +47,7 @@
             }
 
             this.TempData["Message"] = "Blog added successfully.";
-            return this.RedirectToAction("All", "Blogs", new { area = " " });
+            return this.RedirectToAction(nameof(this.All));
         }
 
         public async Task<IActionResult> All(int id = 1)
@@ -68,11 +68,60 @@
             return this.View(model);
         }
 
+        public async Task<IActionResult> Details(int id)
+        {
+            if (!await this.blogService.DoesBlogExistsAsync(id))
+            {
+                return this.RedirectToAction(nameof(this.All));
+            }
+
+            var model = await this.blogService.BlogDetailsByIdAsync<SingleBlogViewModel>(id);
+            return this.View(model);
+        }
+
+        public async Task<IActionResult> Edit(int id)
+        {
+            if (!await this.blogService.DoesBlogExistsAsync(id))
+            {
+                return this.RedirectToAction(nameof(this.All));
+            }
+
+            var model = await this.blogService.BlogDetailsByIdAsync<EditBlogViewModel>(id);
+            return this.View(model);
+        }
+
         [HttpPost]
+        public async Task<IActionResult> Edit(int id, EditBlogViewModel model)
+        {
+            if (!this.ModelState.IsValid)
+            {
+                return this.View(model);
+            }
+
+            var staffId = this.User.FindFirstValue(ClaimTypes.NameIdentifier);
+            try
+            {
+                await this.blogService.EditBlogAsync(model, id, staffId, $"{this.environment.WebRootPath}/images");
+            }
+            catch (Exception)
+            {
+                this.ModelState.AddModelError(string.Empty, "Could not edit this blog");
+                return this.View(model);
+            }
+
+            this.TempData["Message"] = "Blog edited successfully.";
+            return this.RedirectToAction(nameof(this.All));
+        }
+
         public async Task<IActionResult> Delete(int id)
         {
+            if (!await this.blogService.DoesBlogExistsAsync(id))
+            {
+                return this.RedirectToAction(nameof(this.All));
+            }
+
             await this.blogService.DeleteBlogAsync(id);
-            return this.RedirectToAction("All", "Blogs", new { area = "Administration" });
+            return this.RedirectToAction(nameof(this.All));
         }
     }
 }
